+++ tests/testthat/test-ggRelatednessMatrix.R
@@ -3,15 +3,15 @@ test_that("ggRelatednessMatrix returns a gg object", {
   data("redsquirrels")
 
   # Set up the data
-  sumped <- BGmisc::summarizePedigrees(redsquirrels,
-    famID = "famID",
-    personID = "personID",
-    nbiggest = 5
-  )
+#  sumped <- BGmisc::summarizePedigrees(redsquirrels,
+#    famID = "famID",
+#    personID = "personID",
+#    nbiggest = 5
+#  )
 
 
   # Set target family for visualization
-  fam_filter <- sumped$biggest_families$famID[3]
+  fam_filter <- 160 #sumped$biggest_families$famID[3]
 
   # Filter for reasonably sized family, recode sex if needed
   ped_filtered <- redsquirrels %>%
@@ -32,4 +32,64 @@ test_that("ggRelatednessMatrix returns a gg object", {
     )
   )
   expect_s3_class(p_add, "gg")
+  expect_s3_class(p_add, "ggplot")
+  expect_true(is_ggplot(p_add))
+
+  # Check if the plot has the expected title
+  expect_equal(p_add$labels$title, "Additive Genetic Relatedness")
+
+  expect_equal(paste0(p_add$layers[[1]][["constructor"]][[1]]),c("::","ggplot2","geom_tile"))
+  expect_true(p_add$theme$axis.text.x$angle == 90)
+  expect_true(p_add$theme$axis.text.y$angle == 0)
+
 })
+
+test_that("ggRelatednessMatrix handles triangles", {
+  library(BGmisc)
+  data("redsquirrels")
+
+  # Set up the data
+  #  sumped <- BGmisc::summarizePedigrees(redsquirrels,
+  #    famID = "famID",
+  #    personID = "personID",
+  #    nbiggest = 5
+  #  )
+
+
+  # Set target family for visualization
+  fam_filter <- 160 #sumped$biggest_families$famID[3]
+
+  # Filter for reasonably sized family, recode sex if needed
+  ped_filtered <- redsquirrels %>%
+    BGmisc::recodeSex(code_female = "F") %>%
+    dplyr::filter(famID == fam_filter)
+
+  # Calculate relatedness matrices
+  add_mat <- BGmisc::ped2add(ped_filtered, isChild_method = "partialparent", sparse = FALSE)
+
+  p_add <- ggRelatednessMatrix(
+    add_mat,
+    config = list(
+      color_palette = c("white", "orange", "red"),
+      scale_midpoint = 0.55,
+      cluster = FALSE,
+      title = "Additive Genetic Relatedness",
+      text_size = 15,
+      include_upper_triangle = FALSE, # Test upper triangle exclusion
+      include_lower_triangle = TRUE # Test lower triangle inclusion
+    )
+  )
+  expect_s3_class(p_add, "gg")
+  expect_s3_class(p_add, "ggplot")
+  expect_true(is_ggplot(p_add))
+  # Check if the plot has the expected title
+  expect_equal(p_add$labels$title, "Additive Genetic Relatedness")
+  expect_equal(paste0(p_add$layers[[1]][["constructor"]][[1]]),c("::","ggplot2","geom_tile"))
+  expect_true(p_add$theme$axis.text.x$angle == 90)
+  expect_true(p_add$theme$axis.text.y$angle == 0)
+  # Check if the upper triangle is excluded
+  expect_true(all(is.na(p_add$data$value[upper.tri(p_add$data$value)])))
+  # Check if the lower triangle is included
+  expect_true(all(!is.na(p_add$data$value[lower.tri(p_add$data$value)])))
+})
+